Fix `nucypher-porter porter run` redundancy in CLI to instead be `nucypher-porter run`. Use Porter version in REST responses.
