+++ porter/__init__.py
@@ -1,7 +1,6 @@
-from nucypher.__about__ import (
+from porter.__about__ import (
     __author__,  __license__, __summary__, __title__, __version__, __copyright__, __email__, __url__
 )
-
 
 __all__ = [
     "__title__", "__summary__", "__version__", "__author__", "__license__", "__copyright__", "__email__", "__url__"
